start allergy concern act test
